almost Done Back End

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/user"
 import Video from "../models/video"
 
 export const home = async (req, res) => {
@@ -7,7 +8,8 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
     const id = req.params.id;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+    // console.log(video);
     if (!video){
         return res.render("404", {pageTitle: "Video not found."});
     }
@@ -15,20 +17,28 @@ export const watch = async (req, res) => {
 }
 
 export const getEdit = async (req, res) => {
-    const id = req.params.id;
+    const { id } = req.params;
+    const { _id } = req.session.user;
     const video = await Video.findById(id);
     if (!video){
-        return res.render("404", {pageTitle: "Video not found."});
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle: `Editing: ${video.title} `, video});
 }
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { _id } = req.session.user;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({_id : id});
     if (!video){
-        return res.render("404", {pageTitle: "Video not found."});
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -43,22 +53,37 @@ export const getUpload = (req, res) => {
 }
  
 export const postUpload = async (req, res) => {
+    const { _id } = req.session.user;
+    const { path } = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl: path,
+            owner: _id,
             hashtags:Video.formatHashtags(hashtags), 
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch(error){
         console.log(error);
-        return res.render("upload", {pageTitle: "Upload Video", errorMessage: error._message});
+        return res.status(400).render("upload", {pageTitle: "Upload Video", errorMessage: error._message});
     }
 }
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+    const {_id} = req.session.user;
+    const {video} = await Video.findById(id);
+    if(!video){
+        return res.render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id); 
     return res.redirect("/");
 }
